+++ index.js
@@ -1,6 +1,9 @@
-import { Orpheus } from '@roitium/expo-orpheus'
+import { Orpheus, registerOrpheusHeadlessTask } from '@roitium/expo-orpheus'
 import log from './src/utils/log'
-import { finalizeAndRecordCurrentTrack } from './src/utils/player'
+import {
+	finalizeAndRecordCurrentTrack,
+	setDesktopLyrics,
+} from './src/utils/player'
 import toast from './src/utils/toast'
 
 // 定义一个全局变量，避免二次初始化 player
@@ -21,8 +24,10 @@ Orpheus.addListener('onTrackFinished', (event) => {
 	)
 })
 
-Orpheus.addListener('onTrackStarted', (event) => {
-	log.debug('onTrackStarted', event)
+registerOrpheusHeadlessTask(async (event) => {
+	if (event.eventName === 'onTrackStarted') {
+		setDesktopLyrics(event.trackId, event.reason)
+	}
 })
 
 import 'expo-router/entry'